Handle Name Collisions
Why:
When one role in a context shares a method name with the name of another
role, the role is superceded by the method. This clearly leads to
surprising results and can be a rather insidious bug since the class
method being called out of context.

This change addresses the need by:
Detecting name collisions and allowing the user to handle the collision
with a lambda. Three built-in options are offered, or the user can pass in
their own lambda.

	modified:   README.md
	modified:   lib/surrounded/context.rb
	modified:   lib/surrounded/context/name_collision_detector.rb
	modified:   lib/surrounded/context_errors.rb
	modified:   test/name_collisions_test.rb