Handle Name Collisions
Why:
When one role in a context shares a method name with the name of another
role, the role is superceded by the method. This clearly leads to
surprising results and can be a rather insidious bug since the class
method being called out of context.

This change addresses the need by:
Detecting name collisions and allowing the user to handle the collision
with a lambda. Three built-in options are offered, or the user can pass in
their own lambda.

--renamed a method that missed the last commit