NameCollisionDetector now uses Hash.new do block
Why: The NameCollisionDetector#detect_collision method used a silly loop
to set its default values which made the code more complex than necessary.

This change addresses the need by:
Using the Hash.new do block idion reduces the amount of code and the
number of methods required and also localises the default logic to the
point of Hash creation which is easier to follow.