on_name_collision can now accept instance methods
Why: In the previous implementation of
NameCollisionDetector#on_name_collision, the method would only accept a
lambda to handle name collisions. This is restrictive, since the client
may prefer to use a method from the context class.

This change addresses the need by: Altering the
NameCollisionDetectorModule such that the on_name_collision method remains
at a class level but the other methods are automatically injected as
instance methods. By doing so, the class can now accept the name of a
method as a symbol which is then bound to the Context instance and is used
to handle the name collision.

Also tried to revert some of the whitespace removed by my last commit.