Redesigned on_name_collision and refactored NameCollisionHandler
on_name_collision used to accept either a symbol or a block, and then handled the corresponding action with a case statement. This was unwieldy and unnecessarily complex. on_name_collision now only accepts a symbol which is the name of a method in the class which will be called when a name collision is detected.

Also refactored the code to make it more legible and easier to read.